refactor: User 도메인 타입 구조 정리
